Check if workspace exists

+++ apps/remix-ide/src/app/files/workspaceFileProvider.js
@@ -12,7 +12,11 @@ class WorkspaceFileProvider extends FileProvider {
 
     try {
       // make sure "code-sample" has been removed
-      window.remixFileSystem.unlink(this.workspacesPath + '/code-sample')    
+      window.remixFileSystem.exist(this.workspacesPath + '/code-sample').then((exist) => {
+        if (exist) window.remixFileSystem.unlink(this.workspacesPath + '/code-sample').catch((e) => {
+          console.log(e)
+        })
+      })      
     } catch (e) {
       // we don't need to log error if this throws an error
     }
